홈 화면 UI 및 목데이터 연결
